+++ src/main/java/a4/streamx_be/chat/domain/dto/request/AIReqDtoV1.java
@@ -1,3 +1,3 @@
-package a4.streamx_be.domain.dto.request;
+package a4.streamx_be.chat.domain.dto.request;
 
 public record AIReqDtoV1(String prompt, String message) { }
