+++ src/main/java/a4/streamx_be/chat/domain/dto/response/AIResDtoV1.java
@@ -1,0 +1,3 @@
+package a4.streamx_be.chat.domain.dto.response;
+
+public record AIResDtoV1(String message)  { }
