+++ src/main/java/a4/streamx_be/chat/domain/dto/response/AIResDtoV2.java
@@ -1,4 +1,4 @@
-package a4.streamx_be.domain.dto.response;
+package a4.streamx_be.chat.domain.dto.response;
 
 public record AIResDtoV2(String text, String audioUrl) {
 }
