--- src/main/java/a4/streamx_be/domain/dto/response/AIResDtoV1.java
@@ -1,3 +1,0 @@
-package a4.streamx_be.domain.dto.response;
-
-public record AIResDtoV1(String message)  { }
