+++ src/test/java/a4/streamx_be/services/impl/ChatServiceTest.java
@@ -1,7 +1,7 @@
 package a4.streamx_be.services.impl;
 
 import a4.streamx_be.configuration.CharacterConfig;
-import a4.streamx_be.domain.dto.response.AIResDtoV3;
+import a4.streamx_be.chat.domain.dto.response.AIResDtoV3;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,7 +11,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.ai.chat.client.ChatClient;
 import org.springframework.ai.chat.client.advisor.vectorstore.QuestionAnswerAdvisor;
-import org.springframework.ai.chat.model.ChatResponse;
 import org.springframework.ai.chat.model.Generation;
 import org.springframework.ai.openai.OpenAiChatModel;
 import org.springframework.ai.vectorstore.SearchRequest;
